Split up RecordTable
